Clean up and document
